MNIST code
Here is the full code for categorizing the MNIST dataset that I did. I based it off of the tensorflow tutorial, and I tried to comment/annotate it heavily for understanding. There are a few noted areas in which I'll need to go back for better understanding, but this will be the basis for future endeavors!